Add Task module
Tablecloth is missing some sort of async monad primitive (like Task in
Elm, Promise in JavaScript, and Deferred in Jane Street land). This PR
adds a Task module to fill that gap.

API matches https://package.elm-lang.org/packages/elm/core/latest/Task
wherever possible while also keeping to ocaml-isms like the named
argument for ~f.

I've only implemented the bucklescript side for now, but I'd be happy to
do the native side as well if you'd be willing to merge something like
  this. The only caveat is Tablecloth would have to also depend on
  async_kernel to implement Task backed by Deferred, and even the async
  library itself to be able to actually execute the Deferred i.e. to
  implement perform and attempt.

Another option if you want to keep the native deps only on base is to
make this async part into a separate library.

I'll also improve the documentation if this is something you're open to
merging.